English words able to be guessed
HTML was changed to allow for a listener on the on the submit button, the 10000 image allowed to be tracked to main page and findLengthGuesses function added. In js the find length guesses fills in word length and number of guesses based on word pulled. JS also changed to include row id when building table and to use inputGuess() to fill in guessed word in the word table and appropriate images in the image table.